+++ e-app/src/lib.rs
@@ -8,6 +8,7 @@ mod editing;
 mod dialogs;
 mod recent;
 mod git_view;
+mod builtin_completion;
 mod about;
 mod update_view;
 mod updater;